pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2wsgi>=1.10.8",
    "authlib>=1.4.1",
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "flask-jwt-extended>=4.7.1",
    "flask-marshmallow>=1.3.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "marshmallow>=3.26.1",
    "marshmallow-sqlalchemy>=1.4.1",
    "psycopg2-binary>=2.9.10",
    "pycparser>=2.22",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.38",
    "typing-extensions>=4.12.2",
    "uvicorn>=0.34.0",
    "uvicorn-worker>=0.3.0",
    "werkzeug>=3.1.3",
]

[dependency-groups]
dev = [
    "coverage>=7.6.12",
]
